+++ scarletio/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.23'
+__version__ = '1.0.24'
 
 from .core import *
 from .ext import *